refactor(@angular-devkit/build-angular): use base class for esbuild Angular JIT/AOT compilation
Types and functionality that were used by both the AOT and JIT compilation classes within
the esbuild Angular compiler plugin have been extracted into a common base class that
both now extend. This removes duplicate code from both classes.
